fix: use index.ts to repost type.ts
